Support Dll for Windows (#157)
* Refs #13577: install configuration files for dll


* Refs #13577: add dll to ddsrouter_utils


* Refs #13577: dll support for windows


* Refs #13577: Support for common interfaces


* Refs #13577: fix windows github action tool compilation


* Refs #13577: apply changes


* Refs #13577: add fastcdr to core tests


* Refs #13577: uncrustify


* Refs #13577: apply suggestions

+++ ddsrouter_core/src/cpp/reader/implementations/auxiliar/DummyReader.cpp
@@ -59,7 +59,7 @@ utils::ReturnCode DummyReader::take_(
 
     // Move Payload to DDSRouter Payload Pool
     payload_pool_->get_payload(
-        next_data_to_send.payload.size() * sizeof(PayloadUnit),
+        static_cast<uint32_t>(next_data_to_send.payload.size() * sizeof(PayloadUnit)),
         data->payload);
 
     // Set values in Payload as the data was not in the DDSRouter Payload Pool
